+++ packages/contracts/src/support/meta/GetterContext.ts
@@ -1,6 +1,8 @@
 import MetadataContext from "./MetadataContext";
 
 /**
+ * @deprecated Replaced by {@link ClassGetterDecoratorContext}
+ * 
  * Class Getter Decorator Context
  */
 export default interface GetterContext extends ClassGetterDecoratorContext, MetadataContext