+++ packages/contracts/src/support/meta/MethodContext.ts
@@ -1,6 +1,8 @@
 import MetadataContext from "./MetadataContext";
 
 /**
+ * @deprecated Replaced by {@link ClassMethodDecoratorContext}
+ * 
  * Class Method Decorator Context
  */
 export default interface MethodContext extends ClassMethodDecoratorContext, MetadataContext