+++ packages/contracts/src/support/meta/SetterContext.ts
@@ -1,6 +1,8 @@
 import MetadataContext from "./MetadataContext";
 
 /**
+ * @deprecated Replaced by {@link ClassSetterDecoratorContext}
+ * 
  * Class Setter Decorator Context
  */
 export default interface SetterContext extends ClassSetterDecoratorContext, MetadataContext