+++ packages/contracts/src/support/meta/index.ts
@@ -7,6 +7,7 @@ import FieldContext from "./FieldContext";
 import AccessorContext from "./AccessorContext";
 import MetaEntry from "./MetaEntry";
 import MetaTargetContext from "./MetaTargetContext";
+import Kind from "./Kind";
 
 /**
  * Support Meta identifier
@@ -23,6 +24,13 @@ export const SUPPORT_META: unique symbol = Symbol('@aedart/contracts/support/met
  */
 export const METADATA: unique symbol = Symbol.for('metadata');
 
+/**
+ * Symbol used for "target" metadata
+ * 
+ * @type {symbol}
+ */
+export const TARGET_METADATA: unique symbol = Symbol('target_metadata');
+
 export {
     type ClassContext,
     type MethodContext,
@@ -34,6 +42,8 @@ export {
     
     type MetaEntry,
     type MetaTargetContext,
+    
+    Kind
 };
 
 export type * from './types';