+++ packages/contracts/src/support/reflections/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Support Reflections identifier
+ *
+ * @type {Symbol}
+ */
+export const SUPPORT_REFLECTIONS: unique symbol = Symbol('@aedart/contracts/support/reflections');