+++ packages/reflections/src/index.ts
@@ -1,5 +1,0 @@
-import Reflector from "./Reflector";
-
-export {
-    Reflector
-}