+++ packages/support/rollup.config.mjs
@@ -4,6 +4,7 @@ export default createConfig({
     baseDir: new URL('.', import.meta.url),
     external: [
         '@aedart/contracts/support',
-        '@aedart/contracts/support/meta'
+        '@aedart/contracts/support/meta',
+        '@aedart/contracts/support/reflections',
     ]
 });