+++ packages/support/src/reflections/index.ts
@@ -1,0 +1,3 @@
+export * from './isCallable';
+export * from './isClassConstructor';
+export * from './isConstructor';